requirements.txt
django
djangorestframework
django-allauth
django-rest-auth
django-rest-swagger
django-model-utils
factory-boy
